pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "fabricutils"
version = "0.1.0"
description = "Utility functions for Microsoft Fabric notebooks"
authors = [{ name = "Jan Poulsen Skrubbeltrang", email = "jan@skrubbeltrang.net" }]
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
dependencies = [
    "openpyxl",
    "pandas",
    "polars",
    "pyarrow",
    "pydantic",
    "python-dotenv",
    "pyyaml",
]

[dependency-groups]
dev = [
    "duckdb",
    "dummy-notebookutils>=1.5.4",
    "isort",
    "jupyterlab",
    "mimesis>=18.0.0",
    "mypy",
    "pyspark",
    "semantic-link-labs>=0.12.6",
    "semantic-link-sempy>=0.12.2",
]

# Import sorting
[tool.isort]
profile = "ruff"
line_length = 100

# Type checking
[tool.mypy]
python_version = "3.11.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

# Setup
[tool.setuptools.packages.find]
where = ["src"]
